requirements.txt
torch
torchvision
fastapi
uvicorn
matplotlib
pillow
python-multipart